+++ src/lib/prefs/ZoomFactorPref.svelte
@@ -13,10 +13,11 @@
 		class="flex-grow"
 		items={[0.5, 0.75, 1, 1.25, 1.5]}
 		selected={value}
-		onSelectedChangeSingle={(newValue) => {
+		onSelectedChange={(newValue) => {
 			value = newValue;
 			set(newValue);
 		}}
 		getLabel={(percentage) => percentage * 100 + '%'}
+		multiple={false}
 	/>
 </div>